+++ CHANGES.md
@@ -22,6 +22,7 @@
   If the device is trusted, localStorage is used and user data is cached indefinitely.
 - Initial support for XEP-0357 Push Notifications, specifically registering an "App Server".
 - Add support for logging in via OAuth (see the [oauth_providers](https://conversejs.org/docs/html/configurations.html#oauth-providers) setting)
+- XEP-0308: Render message corrections 
 
 ### Bugfixes